requirements.txt
google-api-python-client
google-auth-oauthlib
google-auth-httplib2
yt-dlp
psycopg2-binary
python-dotenv
google-generativeai